+++ packages/fuselage/src/helpers/cssSupports.js
@@ -1,0 +1,3 @@
+import { memoize } from './memoize';
+
+export const cssSupports = memoize((value) => typeof window !== 'undefined' && window.CSS && window.CSS.supports(value));
