+++ packages/fuselage/src/styles/props/position.js
@@ -1,8 +1,9 @@
 import { css } from '@rocket.chat/css-in-js';
 
+import { createLogicalProperties } from '../../helpers/createLogicalProperties';
+import { createPropType } from '../../helpers/createPropType';
+import { cssSupports } from '../../helpers/cssSupports';
 import { memoize } from '../../helpers/memoize';
-import { cssSupports, createLogicalProperties, createPropType } from '../helpers';
-
 
 export const getInsetValue = memoize((propValue) => {
   if (propValue === undefined || propValue === null) {
